requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
google-genai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
structlog>=24.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12
chromadb>=0.5.0
sentence-transformers>=3.0.0
httpx>=0.27.0
langfuse>=2.0.0
python-json-logger>=2.0.7
aiosmtplib>=3.0.0
